requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
redis==5.0.1
aioredis==2.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-dotenv==1.0.0
orjson==3.9.10
structlog==24.1.0
prometheus-client==0.19.0
pytest==7.4.4
pytest-asyncio==0.23.3